adding contact extensions.
