+++ iOS/Flows/Channels/Cells/ReservationCell.swift
@@ -7,10 +7,13 @@
 //
 
 import Foundation
+import Combine
+import TMROLocalization
 
 class ReservationCell: CollectionViewManagerCell {
 
     let button = Button()
+    private var cancellables = Set<AnyCancellable>()
 
     override func initializeSubviews() {
         super.initializeSubviews()
@@ -21,9 +24,19 @@ class ReservationCell: CollectionViewManagerCell {
     func configure(with reservation: Reservation) {
 
         if let contactId = reservation.contactId {
-
+            ContactsManger.shared.searchForContact(with: .identifier(contactId))
+                .mainSink { (result) in
+                    switch result {
+                    case .success(let contacts):
+                        guard let contact = contacts.first, let name = contact.fullName else { return }
+                        let text = LocalizedString(id: "", arguments: [name], default: "Remind @(name)?")
+                        self.button.set(style: .normal(color: .purple, text: text))
+                    case .error(_):
+                        break
+                    }
+                }.store(in: &self.cancellables)
         } else {
-            self.button.set(style: .normal(color: .purple, text: "Send RSVP"))
+            self.button.set(style: .icon(image: UIImage(systemName: "person.badge.plus")!))
         }
 
         self.contentView.layoutNow()
